Rough Prototype of Auditory Covid 19 Self Assessment
Two functions, one for simple yes/no answers and one for emergency yes/no answers, exist for the questionnaire. Google Speech API picks up on "yes" and "no" and reacts accordingly. For simple questions, an "unsafe" answer changes the final result to a stay-at-home order. For emergency questions, an undesirable answer leads to advice to call 911 and the program is stopped. To communicate with the user, pyttsx 3 is used to convert text to speech.